pyproject.toml
[tool.poetry]
name = "ml-dengue-predict"
version = "0.1.0"
description = ""
authors = ["Sandro Loch <es.loch@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fetchinfodenguedata"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.3"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.20"
psycopg2 = "^2.9.7"
jupyterlab = "^4.0.5"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.2"
numpy = "^1.25.2"
graphviz = "^0.20.1"
pylatex = "^1.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
